Remove SetProperty(...cPlayer)
Enchantment table, anvil windows are already opened one per-player.

+++ src/ClientHandle.h
@@ -233,7 +233,7 @@ public:  // tolua_export
 	void SendWholeInventory             (const cWindow & a_Window);
 	void SendWindowClose                (const cWindow & a_Window);
 	void SendWindowOpen                 (const cWindow & a_Window);
-	void SendWindowProperty             (const cWindow & a_Window, short a_Property, short a_Value);
+	void SendWindowProperty             (const cWindow & a_Window, size_t a_Property, short a_Value);
 
 	// tolua_begin
 	const AString & GetUsername(void) const;
